cart storing added, no delete from cart

+++ src/main/java/com/danila/market/service/ProductService.java
@@ -1,8 +1,11 @@
 package com.danila.market.service;
 
 import com.danila.market.entity.Product;
+import com.danila.market.exceptions.AppError;
 import com.danila.market.repository.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -13,17 +16,20 @@ import java.util.List;
 public class ProductService {
     private final ProductRepository productRepository;
 
-    public ResponseEntity<?> getAllProducts() {
-        List<Product> products = productRepository.findAll();
-        return ResponseEntity.ok(products);
+    public List<Product> getAllProducts() {
+        var products = productRepository.findAll();
+        return products;
     }
 
-    public ResponseEntity<?> saveNewProduct(Product product) {
-        productRepository.save(product);
-        return ResponseEntity.ok(product);
+    public Product saveNewProduct(Product product) {
+        var savedProduct = productRepository.save(product);
+        return savedProduct;
     }
-    public ResponseEntity<?> deleteProductById(int id) {
-        productRepository.deleteById(id);
-        return ResponseEntity.ok(id);
+    public void deleteProductById(int id) {
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Product with id " + id + " not found");
+        }
     }
 }
